requirements.txt
absl-py==0.4.0
adium-theme-ubuntu==0.3.4
alembic==1.0.5
argcomplete==1.8.2
astor==0.7.1
attrs==15.2.0
autobahn==0.10.3
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
bcrypt==3.1.5
beautifulsoup4==4.5.3
bzr==2.7.0
cached-property==1.3.0
catkin-pkg==0.3.9
catkin-pkg-modules==0.3.9
certifi==2018.11.29
cffi==1.11.5
chardet==2.3.0
Click==7.0
cloudpickle==0.5.3
configobj==5.0.6
crcmod==1.7
cryptography==1.7.2
cycler==0.10.0
dask==0.18.2
DateTime==4.3
decorator==4.3.0
defusedxml==0.4.1
docker-compose==1.8.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
docx2txt==0.6
EbookLib==0.15
ecdsa==0.13
empy==3.3.2
enum34==1.1.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gast==0.2.0
grpcio==1.14.1
httplib2==0.9.1
idna==2.8
imageflow==0.0.3
imgaug==0.2.6
ipaddress==1.0.16
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.5.1
keyring==7.3
kiwisolver==1.0.1
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lxml==4.3.0
lz4==0.7.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.2.3
mercurial==3.7.3
mock==2.0.0
mpi4py==1.3.1
msgpack-python==0.4.6
ndg-httpsclient==0.4.0
netifaces==0.10.4
networkx==2.1
nltk==3.3
nose==1.3.7
numpy==1.11.0
oauth==1.0.1
opencv-python==4.0.0.21
PAM==0.4.2
paramiko==1.16.0
pbr==4.2.0
pdf2image==1.4.0
Pillow==5.2.0
pocketsphinx==0.1.3
protobuf==3.6.1
psutil==3.4.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.0.29
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
PyJWT==1.7.1
PyOpenGL==3.0.2
pyparsing==2.0.3
PyPDF2==1.26.0
pyserial==3.0.1
Pyste==0.9.10
pytesseract==0.2.6
python-apt==1.1.0b1
python-dateutil==2.4.2
python-editor==1.0.3
python-pptx==0.6.5
python-snappy==0.5
pytz==2014.10
PyWavelets==0.5.2
PyYAML==3.11
requests==2.21.0
roman==2.0.0
rosdep==0.11.8
rosdistro==0.6.4
rosdistro-modules==0.6.4
rosinstall==0.7.8
rosinstall-generator==0.1.13
rospkg==1.1.4
rospkg-modules==1.1.4
scikit-image==0.14.0
SecretStorage==2.1.3
service-identity==16.0.0
simplejson==3.8.1
six==1.10.0
SpeechRecognition==3.6.3
SQLAlchemy==1.2.15
subprocess32==3.5.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
textract==1.6.1
texttable==0.8.1
toolz==0.9.0
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
typing==3.5.3.0
unity-lens-photos==1.0
urllib3==1.24.1
vcstools==0.1.40
wadllib==1.3.2
websocket-client==0.18.0
Werkzeug==0.14.1
wstool==0.1.17
xlrd==1.0.0
XlsxWriter==1.1.2
zope.interface==4.1.3
